2024/Character Portrait Generator/requirements.txt
gradio
Pillow
numpy
websocket-client
